Extracts Classroom Editor into its own Component
