+++ src/components/Classroom.tsx
@@ -1,5 +1,6 @@
-import { FC, CSSProperties, useState } from "react";
+import { FC, CSSProperties } from "react";
 import { Square } from "./Square";
+import { ClassroomEditor } from "./ClassroomEditor";
 import { Room } from "../lib/Room";
 
 export interface ClassroomProps {
@@ -13,27 +14,12 @@ const boardStyles: CSSProperties = {
   flexWrap: "wrap",
 };
 
-const asideStyles: CSSProperties = {
-  width: "25%",
-  height: "100%",
-  display: "flex",
-  flexDirection: "column",
-  justifyContent: "space-between",
-};
-
 export const Classroom: FC<ClassroomProps> = ({ room }) => {
-  const [studentName, setStudentName] = useState("");
-
   const numberOfSquares = 100;
   const rootNumberOfSquares = Math.sqrt(numberOfSquares);
   const squareStyles: CSSProperties = {
     width: `${rootNumberOfSquares}%`,
     height: `${rootNumberOfSquares}%`,
-  };
-
-  const addStudent = () => {
-    room.assignNewStudent(studentName);
-    setStudentName("");
   };
 
   return (
@@ -47,21 +33,7 @@ export const Classroom: FC<ClassroomProps> = ({ room }) => {
           );
         })}
       </div>
-      <aside style={asideStyles}>
-        <h2>Editor</h2>
-        <h3>Tisch Anordnung</h3>
-        <button onClick={() => room.generateTablePreset(3)}>
-          Parallele Reihen
-        </button>
-        <label htmlFor="add-student">Neuer Schüler</label>
-        <input
-          id="add-student"
-          type="text"
-          value={studentName}
-          onChange={(e) => setStudentName(e.target.value)}
-        />
-        <button onClick={addStudent}>+</button>
-      </aside>
+      <ClassroomEditor room={room} />
     </div>
   );
 };
