+++ src/components/ClassroomEditor.tsx
@@ -1,0 +1,63 @@
+import { CSSProperties, FC, useState } from "react";
+import { Room } from "../lib/Room";
+
+interface ClassroomEditorProps {
+  room: Room;
+}
+
+const asideStyles: CSSProperties = {
+  width: "25%",
+  minHeight: "100%",
+};
+
+const asideBlockStyles: CSSProperties = {
+  borderRadius: "4px",
+  border: "1px solid lightgrey",
+  padding: "8px",
+  margin: "12px 0px",
+};
+
+export const ClassroomEditor: FC<ClassroomEditorProps> = ({ room }) => {
+  const [studentName, setStudentName] = useState("");
+
+  const addStudent = () => {
+    room.assignNewStudent(studentName);
+    setStudentName("");
+  };
+
+  return (
+    <aside style={asideStyles}>
+      <h2>Editor</h2>
+      <div style={asideBlockStyles}>
+        <h3>Vorlage</h3>
+        <button onClick={() => room.generateTablePreset(3)}>
+          Parallele Reihen
+        </button>
+      </div>
+      <div style={asideBlockStyles}>
+        <label htmlFor="add-student">
+          <h3>Neuer Schüler</h3>
+        </label>
+        <input
+          id="add-student"
+          type="text"
+          value={studentName}
+          onChange={(e) => setStudentName(e.target.value)}
+          style={{ width: "100%" }}
+        />
+        <button
+          onClick={addStudent}
+          style={{ width: "100%", marginTop: "8px" }}
+        >
+          +
+        </button>
+      </div>
+      <div style={asideBlockStyles}>
+        <h3>Neuen Tisch erstellen </h3>
+        <p>
+          <em>Doppelklick auf ein freies Feld</em>
+        </p>
+      </div>
+    </aside>
+  );
+};
